requirements.txt
##############################
# CORE BACKEND
##############################
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1

pydantic==2.5.0
pydantic-settings==2.1.0

##############################
# AI / MACHINE LEARNING
##############################
insightface==0.7.3
onnxruntime==1.17.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
scikit-learn==1.3.2
imagehash==4.3.1

sentence-transformers==2.2.2
chromadb==0.4.22

##############################
# MEDIA / METADATA PROCESSING
##############################
ffmpeg-python==0.2.0

# Removed py3exiv2 — NOT Windows compatible and version 0.10.0 does not exist
# piexif handles EXIF read/write on Windows
piexif==1.1.3

python-magic-bin==0.4.14

##############################
# WATCHDOG / TASKS
##############################
watchdog==3.0.0
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6

##############################
# UTILITIES
##############################
httpx==0.27.0
send2trash==1.8.2
pytesseract==0.3.10
cryptography==41.0.7
timezonefinder==6.2.0
geopy==2.4.1
psutil==5.9.7

##############################
# WINDOWS-SPECIFIC
##############################
pywin32==306
winshell==0.6

##############################
# OPTIONAL — If you have NVIDIA GPU
##############################
# torch==2.1.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision==0.16.0+cu118 --index-url https://download.pytorch.org/whl/cu118
